# packages/api-client/openapi.json
{
  "components": {
    "schemas": {
...
      "BackupListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/BackupManifest"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "BackupListResponse",
        "type": "object"
      },
      "BackupManifest": {
        "additionalProperties": true,
        "properties": {
          "completed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "conversation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conversation Id"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "operations": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Operations",
            "type": "array"
          },
          "state": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BackupState"
              },
              {
                "type": "null"
              }
            ]
          },
          "timestamp": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timestamp"
          }
        },
        "required": [
          "id"
        ],
        "title": "BackupManifest",
        "type": "object"
      },
      "BackupState": {
        "enum": [
          "active",
          "completed",
          "abandoned",
          "quarantined"
        ],
        "title": "BackupState",
        "type": "string"
      },
...
      "ProblemDetails": {
        "additionalProperties": true,
        "description": "RFC 9457-compatible public error contract.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "requestId": {
            "title": "Requestid",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "code",
          "detail",
          "requestId"
        ],
        "title": "ProblemDetails",
        "type": "object"
      },
...
      "RollbackRequest": {
        "properties": {
          "backupGroupId": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Backupgroupid",
            "type": "string"
          },
          "operationIndex": {
            "minimum": 0.0,
            "title": "Operationindex",
            "type": "integer"
          }
        },
        "required": [
          "backupGroupId",
          "operationIndex"
        ],
        "title": "RollbackRequest",
        "type": "object"
      },
      "RollbackResponse": {
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          },
          "success": {
            "const": true,
            "default": true,
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "message"
        ],
        "title": "RollbackResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "description": "Deployment access token for non-browser clients.",
        "scheme": "bearer",
        "type": "http"
      },
      "sessionCookie": {
        "description": "Signed HttpOnly cookie issued by the same-origin POST exchange.",
        "in": "cookie",
        "name": "markinote_access",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "summary": "Typed API for the MarkiNote document workspace and AI agent",
    "title": "MarkiNote",
    "version": "4.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/operations/backups": {
      "get": {
        "operationId": "list_backups_api_v1_operations_backups_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BackupListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Invalid request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Authentication required"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Request forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Resource conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Request or resource too large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Contract validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "sessionCookie": []
          }
        ],
        "summary": "List Backups",
        "tags": [
          "operations"
        ]
      }
    },
    "/api/v1/operations/rollback": {
      "post": {
        "operationId": "rollback_operation_api_v1_operations_rollback_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RollbackRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RollbackResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Invalid request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Authentication required"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Request forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Resource conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Request or resource too large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Contract validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "sessionCookie": []
          }
        ],
        "summary": "Rollback Operation",
        "tags": [
          "operations"
        ]
      }
    },
...
  }
}